backend/requirements.txt
# Claims Processing Platform - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Caching
redis==5.0.1

# AWS/Cloud Services
boto3==1.34.25

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
factory-boy==3.3.0

# Logging & Monitoring
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Code Quality
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
